change project structure for Module
