+++ src/EDApiConnector/connector/base/exception.py
