+++ src/EDApiConnector/connector/edsm/marketApi.py
@@ -1,5 +1,5 @@
-from connector.base import exception
-from connector.base.base import ApiEntryPoint
+from EDApiConnector.connector.base import exception
+from EDApiConnector.connector.base.base import ApiEntryPoint
 
 
 class Market(ApiEntryPoint):
